requirements.txt
pvporcupine==3.0.2
pyaudio==0.2.14
requests==2.31.0
pyyaml==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
aiofiles==23.2.1
wave